agents/borkov/pyproject.toml
[project]
name = "pr-report-coral-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "langchain==0.3.25",
    "langchain-community==0.3.24",
    "langchain-experimental==0.3.4",
    "langchain-groq==0.3.4",
    "langchain-mcp-adapters==0.1.7",
    "langchain-openai==0.3.26",
    "langchain-google-genai>=2.0.0",
    "feedparser>=6.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "reportlab>=4.0.0",
    "google-cloud-storage>=2.10.0",
    "uv>=0.7.17",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
]
